LPS-83949 Move JSLoaderModulesTracker to its own package; no logic changes
