+++ modules/apps/frontend-js/frontend-js-loader-modules-extender/src/test/java/com/liferay/frontend/js/loader/modules/extender/internal/JSLoaderModulesServletTest.java
@@ -18,6 +18,7 @@ import aQute.bnd.osgi.Constants;
 
 import aQute.lib.converter.Converter;
 
+import com.liferay.frontend.js.loader.modules.extender.internal.config.generator.JSLoaderModulesTracker;
 import com.liferay.frontend.js.loader.modules.extender.internal.npm.NPMRegistryImpl;
 import com.liferay.frontend.js.loader.modules.extender.npm.NPMRegistry;
 import com.liferay.petra.string.StringPool;
